refactor(#142/*): place success code global/response 로 이동
