+++ src/main/java/org/sopt/snappinserver/global/response/code/place/PlaceSuccessCode.java
@@ -1,4 +1,4 @@
-package org.sopt.snappinserver.api.v1.place.code;
+package org.sopt.snappinserver.global.response.code.place;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
